+++ client/src/io/worxbend/gitea4s/http/IssueCommentListParams.scala
@@ -1,0 +1,21 @@
+package io.worxbend.gitea4s.http
+
+import java.time.Instant
+
+final case class IssueCommentListParams(
+    since: Option[Instant] = None,
+    before: Option[Instant] = None
+)
+
+object IssueCommentListParams:
+  val default: IssueCommentListParams = IssueCommentListParams()
+
+final case class RepositoryCommentListParams(
+    since: Option[Instant] = None,
+    before: Option[Instant] = None,
+    page: Option[Int] = None,
+    limit: Option[Int] = None
+)
+
+object RepositoryCommentListParams:
+  val default: RepositoryCommentListParams = RepositoryCommentListParams()